requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.18.0
elevenlabs>=0.3.0
click>=8.1.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-vcr>=1.0.0
